refactor: add marker to extension fn
This prevents mistakes when using a extension-fn factory. Before this change under certain circumstances it was possible to not call the factory fn without getting errors on both design and runtime! This happened in all cases where the factory function was assignable to the ExtensionFn type. That's why we made the extension fn type different from "normal" functions via the marker, effectively preventing the drawn scenarios